Introduce release build types, introduce sdk flavors
millhouse and calrissian are creating two seprarate apps in the market and
should be build types and not flavors. Introduced two flavors, support and
kitkat. Support includes support for the default minimum api of 8 and kitkat is
set with a minimum of 19. Note that converting the millhouse/calrissian
flavors to buildTypes was pretty straight forward, aided by the source
directories for the build and the flavor being equal,
src/<build/variant-name> .