Create and release APK on Google Play Alpha Channel.
Updated the code to include some images needed for Google Play profile.
The gradle file has been udated. When the command assembleRelease is
called from the command line the millhouse, release, and calrissian apks
will be creeated with the aid of the right information. During the build,
if System.console is unavailable, the process will not be started.
The apks have been uploaded to the google play alpha. One may get and down
load the apk by going to
https://groups.google.com/forum/#!forum/icndbandroidclient
and becoming affiliated.