Include debug and release build types, two flavors
updated build.gradle to build both debug and release for calrissian and
willson build types. Included the button and the slider activity flavors.
The slider activity will include a slider with an image instead of a
button to get the next joke.
Removed some files in the way of the build.