Include a Joke content provider, MainActivity can use it.
Inlcuded a ContentProvider subclass for storring jokes in a local
SQLiteDatabase, using SQLiteOpenHelper. The Provider also includes a
Contract class for organizing the table names. There is  an inner
class to JokeProvider, DBSchema which defines the SQLite statements
that must be used to manage the database beyond crud ops.
MainActivity has a new getJoke() method which uses JokeRetriever to
retreive a joke using itself as the JokeResponder where the db
insert happens. An integration test checks the functionality of
the method and its impacked on the db.