JokeViewerFragment now implements LoaderCallbacks
The JokeViewerFragment implements LoaderCallbacks for Cursor. The
implementation creates an instance of CursorLoader that watches the Joke
Provider table URI. When the table is updated, the adapter is updated with
the new cursor, moving the ViewPager position to the last position of any
added Jokes. During onStop() the position is saved and is set to the
current view during a restart of the app if there is no saved instance.
Thus, position is preserved over instances.