JokeViewerFragment now gets it's own jokes.
JokeViewerFragment now recognizes when there are no jokes in db and gets
one. if the user slides far to the right the JokeViewerFragment will
add a new fragment to the adapter for the next possible db id. The
fragment will then get the item and update itself.
Renamed a flavor and folders to JokeViewerFlaver